+++ tests/tests_pytorch/loops/test_fetchers.py
@@ -168,7 +168,7 @@ DATASET_LEN = 64
 
 
 @pytest.mark.parametrize("automatic_optimization", [False, True])
-def test_fetching_dataloader_iter_opt(automatic_optimization, tmpdir):
+def test_fetching_dataloader_iter_opt(automatic_optimization, tmp_path):
     class TestModel(BoringModel):
         def __init__(self, *args, automatic_optimization: bool = False, **kwargs):
             super().__init__(*args, **kwargs)
@@ -207,7 +207,7 @@ def test_fetching_dataloader_iter_opt(automatic_optimization, tmpdir):
             assert self.count == 64
 
     model = TestModel(automatic_optimization=automatic_optimization)
-    trainer = Trainer(default_root_dir=tmpdir, max_epochs=1, accelerator="cpu")
+    trainer = Trainer(default_root_dir=tmp_path, max_epochs=1, accelerator="cpu")
     trainer.fit(model)
 
 
@@ -290,9 +290,9 @@ class AsyncBoringModel(BoringModel):
         return DataLoader(RandomDataset(BATCH_SIZE, DATASET_LEN))
 
 
-def test_training_step_with_dataloader_iter(tmpdir) -> None:
+def test_training_step_with_dataloader_iter(tmp_path) -> None:
     """A baseline functional test for `training_step` with dataloader access."""
-    trainer = Trainer(max_epochs=1, default_root_dir=tmpdir, accelerator="cpu")
+    trainer = Trainer(max_epochs=1, default_root_dir=tmp_path, accelerator="cpu")
     m = AsyncBoringModel()
     trainer.fit(m)
     assert m.num_batches_processed == DATASET_LEN, f"Expect all {DATASET_LEN} batches to be processed."
@@ -406,7 +406,7 @@ def test_step_methods_with_dataloader_iter(limit_sanity_val_batches, limit_train
 
 
 @pytest.mark.parametrize("trigger_stop_iteration", [False, True])
-def test_stop_iteration_with_dataloader_iter(trigger_stop_iteration, tmpdir):
+def test_stop_iteration_with_dataloader_iter(trigger_stop_iteration, tmp_path):
     """Verify that StopIteration properly terminates the training when this is triggered from the current
     `dataloader_iter`"""
     EXPECT_NUM_BATCHES_PROCESSED = 2
@@ -428,7 +428,7 @@ def test_stop_iteration_with_dataloader_iter(trigger_stop_iteration, tmpdir):
                 return DataLoader(RandomDataset(BATCH_SIZE, 2 * EXPECT_NUM_BATCHES_PROCESSED))
             return DataLoader(RandomDataset(BATCH_SIZE, EXPECT_NUM_BATCHES_PROCESSED))
 
-    trainer = Trainer(max_epochs=1, default_root_dir=tmpdir, accelerator="cpu")
+    trainer = Trainer(max_epochs=1, default_root_dir=tmp_path, accelerator="cpu")
     m = TestModel(trigger_stop_iteration)
     trainer.fit(m)
     expected = EXPECT_NUM_BATCHES_PROCESSED
@@ -437,7 +437,7 @@ def test_stop_iteration_with_dataloader_iter(trigger_stop_iteration, tmpdir):
     assert m.num_batches_processed == expected
 
 
-def test_transfer_hooks_with_unpacking(tmpdir):
+def test_transfer_hooks_with_unpacking(tmp_path):
     """This test asserts the `transfer_batch` hooks are called only once per batch."""
 
     class RandomDictDataset(RandomDataset):
@@ -476,7 +476,7 @@ def test_transfer_hooks_with_unpacking(tmpdir):
             x, _ = batch
             return super().validation_step(x, batch_idx)
 
-    trainer = Trainer(default_root_dir=tmpdir, max_epochs=1, num_sanity_val_steps=0)
+    trainer = Trainer(default_root_dir=tmp_path, max_epochs=1, num_sanity_val_steps=0)
     dm = BoringDataModule()
     trainer.fit(TestModel(), datamodule=dm)
     assert dm.count_called_on_before_batch_transfer == 4